remove unnecessary loc in the js code
